Refactoring for file names and folder structure
