+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { Navigate, Route, Routes, useParams } from "react-router-dom";
-import Result from "./pages/result/Result";
+import Results from "./pages/play/Results";
 import Home from "./pages/home/Home";
-import QuizTyping from "./pages/typing-quiz/QuizTyping";
+import Typing from "./pages/play/quizes/Typing";
 import Header from "./components/header/Header";
 import { getQuiz } from "./utils/getQuiz";
 import ResultsProvider from "./results-provider";
@@ -12,7 +12,7 @@ function App() {
     const quiz = getQuiz(quizName);
 
     if (quiz) {
-      return <QuizTyping quiz={quiz} />;
+      return <Typing quiz={quiz} />;
     } else if (!quiz) {
       return <Navigate to="/" replace />;
     }
@@ -25,7 +25,7 @@ function App() {
         <ResultsProvider>
           <Routes>
             <Route path="/" element={<Home />} />
-            <Route path="/result" element={<Result />} />
+            <Route path="/results" element={<Results />} />
             <Route path="/play">
               {/* Add more routes here as I build more types of quizes */}
               <Route path="typing/:quizName" element={<FindQuiz />} />
