--- src/components/CountdownTimer/CountdownTimer.css
@@ -1,14 +1,0 @@
-.countdownTimer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.countdownTimer__icon {
-  width: 50px !important;
-  height: 50px !important;
-}
-
-.countdownTimer__text {
-  font-size: 35px;
-}
